+++ frontend/srcs/src/app/modules/shared/base/base.component.ts
@@ -4,8 +4,6 @@ import { lastValueFrom } from 'rxjs';
 import { ApiService } from '../../../services';
 import { AlertModel, ApiModel } from '../../../models';
 import { environment } from 'src/environments/environment';
-// import { UriConstants } from '../../../utils';
-
 
 
 interface KeyParams {